Without pagination

+++ app/Http/Controllers/ExemplarController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Brand;
 use App\Exemplar;
+use App\Product;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
@@ -26,12 +27,26 @@ class ExemplarController extends Controller
     public function created( Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'name' => 'unique:exemplars|required|min:3|max:50',
+            'name' => 'required|max:50',
         ]);
 
-        if ($validator->fails()) {
+        $exemplar_ = Exemplar::where( 'name',$request->get('name') )->first();
+        $name="";
+
+        if( strlen( $request->get('name') ) <4 )
+            $name = "errorName";
+
+        if ( $validator->fails() OR $name == "errorName" OR $exemplar_ != null )
+        {
             $data['errors'] = $validator->errors();
-            return redirect('modelo')
+
+            if( $name == "errorName" )
+                $data['errors']->add("name", "El nombre del modelo  debe tener por lo menos 3 caracteres ");
+            else if( $exemplar_ != null )
+                $data['errors']->add("name", "No puede registrar 2  modelos con el mismo nombre");
+
+            return redirect('modelo/registrar')
+                ->withInput($request->all())
                 ->with($data);
         }
 
@@ -42,7 +57,6 @@ class ExemplarController extends Controller
         ]);
 
         $exemplar->save();
-
 
         return redirect('modelo');
     }
@@ -56,12 +70,32 @@ class ExemplarController extends Controller
     public function edit( Request $request )
     {
         $validator = Validator::make($request->all(), [
-            'name' => 'unique:exemplars|required|min:3|max:50'
+            'name' => 'required|max:50'
         ]);
 
-        if ($validator->fails()) {
+        $exemplar_ = Exemplar::where( 'name',$request->get('name') )->first();
+        $name="";$exemplar_id  ="";$exemplar_repeated="";
+
+        if( $exemplar_ != null )
+        {
+            $exemplar_id = $exemplar_->id;
+            if( $exemplar_id != $request->get('id') )
+                $exemplar_repeated = "errorRepeated";
+        }
+        if( strlen( $request->get('name') ) <4 )
+            $name = "errorName";
+
+        if ( $validator->fails() OR $name == "errorName" OR $exemplar_repeated != null )
+        {
             $data['errors'] = $validator->errors();
+
+            if( $name == "errorName" )
+                $data['errors']->add("name", "El nombre del modelo debe tener por lo menos 3 caracteres ");
+            else if( $exemplar_repeated != null )
+                $data['errors']->add("name", "No puede registrar 2  modelos con el mismo nombre");
+
             return redirect('modelo')
+                ->withInput($request->all())
                 ->with($data);
         }
 
@@ -78,6 +112,19 @@ class ExemplarController extends Controller
         $validator = Validator::make($request->all(), [
             'id' => 'exists:exemplars,id'
         ]);
+
+        $product = Product::where('exemplar_id',$request->get('id'))->first();
+
+        if ($validator->fails() OR $product != null)
+        {
+            $data['errors'] = $validator->errors();
+            if( $product != null )
+                $data['errors']->add("id", "No puede eliminar el modelo seleccionada, porque existen productos que dependen de él.");
+
+            return redirect('modelo')
+                ->withInput($request->all())
+                ->with($data);
+        }
 
         $exemplar = Exemplar::find( $request->get('id') );
         $exemplar->delete();
